# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/karyawans": {
            "get": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "List all karyawan",
                "description": "Menampilkan daftar semua karyawan",
                "operationId": "69e45aa66e4f8c21f68a4be812d9ccc8",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data karyawan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Karyawan"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Create new karyawan",
                "description": "Membuat data karyawan baru",
                "operationId": "37325a7e80370625466df60414cdbd71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "departemen_id",
                                    "nama",
                                    "jabatan",
                                    "gaji_pokok",
                                    "email"
                                ],
                                "properties": {
                                    "departemen_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nama": {
                                        "type": "string",
                                        "example": "Lis"
                                    },
                                    "jabatan": {
                                        "type": "string",
                                        "example": "Manager"
                                    },
                                    "gaji_pokok": {
                                        "type": "number",
                                        "example": "8.00"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "lis@gmail.com"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "kober, purwokerto"
                                    },
                                    "no_telepon": {
                                        "type": "string",
                                        "example": "0821323545122"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil membuat karyawan baru",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Karyawan"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/karyawans/{karyawan}": {
            "get": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Show a specific karyawan",
                "description": "Menampilkan detail karyawan berdasarkan ID",
                "operationId": "fe4a2c07d435f279e95b2586db0d6389",
                "parameters": [
                    {
                        "name": "karyawan",
                        "in": "path",
                        "description": "ID karyawan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail data karyawan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Karyawan"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Karyawan not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Update a specific karyawan",
                "description": "Memperbarui data karyawan berdasarkan ID",
                "operationId": "534232d060c5ce047be33fc67a91171c",
                "parameters": [
                    {
                        "name": "karyawan",
                        "in": "path",
                        "description": "ID karyawan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "departemen_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nama": {
                                        "type": "string",
                                        "example": "Lis"
                                    },
                                    "jabatan": {
                                        "type": "string",
                                        "example": "Manager"
                                    },
                                    "gaji_pokok": {
                                        "type": "number",
                                        "example": "8.00"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "lis@gmail.com"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "kober, purwokerto"
                                    },
                                    "no_telepon": {
                                        "type": "string",
                                        "example": "0821323545122"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil memperbarui karyawan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Karyawan"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Delete a specific karyawan",
                "description": "Menghapus data karyawan berdasarkan ID",
                "operationId": "57b3b01f359725a62207d8c0feb3a9a5",
                "parameters": [
                    {
                        "name": "karyawan",
                        "in": "path",
                        "description": "ID karyawan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil menghapus karyawan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/karyawans/departemen/{departemen_id}": {
            "get": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Get karyawan by departemen",
                "description": "Menampilkan daftar karyawan berdasarkan ID departemen",
                "operationId": "c62078d755c7456616647f6b22fd1a66",
                "parameters": [
                    {
                        "name": "departemen_id",
                        "in": "path",
                        "description": "ID departemen",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan data karyawan berdasarkan departemen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Karyawan"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Departemen not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/karyawans/search/{nama}": {
            "get": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Search karyawan by name",
                "description": "Mencari karyawan berdasarkan nama",
                "operationId": "365a9f38c2de173c0b3f98db7684e727",
                "parameters": [
                    {
                        "name": "nama",
                        "in": "path",
                        "description": "Nama karyawan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menemukan karyawan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Karyawan"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Karyawan not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Karyawan": {
                "required": [
                    "nama",
                    "jabatan",
                    "gaji_pokok",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "jabatan": {
                        "type": "string",
                        "example": "Manager"
                    },
                    "gaji_pokok": {
                        "type": "number",
                        "format": "float",
                        "example": 5000000
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-03T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-03T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication using Laravel Sanctum",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "security": [
                {
                    "sanctum": []
                }
            ]
        }
    },
...
}